refactor: split hydration functions from utility functions
Split utility fuctions and hydration functions into separate directories. Updated paths and created
a "german" group of paths in the tsconfig.
