+++ apps/db_management/src/models/german/README.md
@@ -1,0 +1,5 @@
+### To do:
+
+- https://german.yabla.com/lesson-The-verb-gelingen-1188
+- add impersonal flag, see above ^^^
+- add dative flag
