+++ apps/db_management/src/models/german/hydrationFunctions/irregularPartizipConjugation.spec.ts
